requirements.txt
pyautogui>=0.9.53
Pillow>=10.1.0
pynput>=1.7.6
numpy>=1.24.0
